+++ ocean_lib/assets/asset.py
@@ -5,11 +5,10 @@
 import copy
 import json
 import logging
-from typing import List, Optional
+from typing import Optional
 
 from enforce_typing import enforce_types
 
-from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.assets.credentials import AddressCredential
 from ocean_lib.services.service import Service
 from ocean_lib.utils.utilities import create_checksum
@@ -193,64 +192,6 @@ class Asset(AddressCredential):
         )
 
     @enforce_types
-    def remove_publisher_trusted_algorithm(
-        self, compute_service: Service, algo_did: str
-    ) -> list:
-        """Returns a trusted algorithms list after removal."""
-        trusted_algorithms = compute_service.get_trusted_algorithms()
-        if not trusted_algorithms:
-            raise ValueError(
-                f"Algorithm {algo_did} is not in trusted algorithms of this asset."
-            )
-        trusted_algorithms = [ta for ta in trusted_algorithms if ta["did"] != algo_did]
-        trusted_algo_publishers = compute_service.get_trusted_algorithm_publishers()
-        self.update_compute_values(
-            trusted_algorithms,
-            trusted_algo_publishers,
-            allow_network_access=True,
-            allow_raw_algorithm=False,
-        )
-        assert (
-            self.get_service("compute").compute_values["publisherTrustedAlgorithms"]
-            == trusted_algorithms
-        ), "New trusted algorithm was not removed. Failed when updating the list of trusted algorithms. "
-
-        return trusted_algorithms
-
-    @enforce_types
-    def remove_publisher_trusted_algorithm_publisher(
-        self, compute_service: Service, publisher_address: str
-    ) -> list:
-        """
-        :return: List of trusted algo publishers not containing `publisher_address`.
-        """
-
-        trusted_algorithm_publishers = [
-            tp.lower() for tp in compute_service.get_trusted_algorithm_publishers()
-        ]
-        publisher_address = publisher_address.lower()
-        if not trusted_algorithm_publishers:
-            raise ValueError(
-                f"Publisher {publisher_address} is not in trusted algorithm publishers of this asset."
-            )
-
-        trusted_algorithm_publishers = [
-            tp for tp in trusted_algorithm_publishers if tp != publisher_address
-        ]
-        trusted_algorithms = compute_service.get_trusted_algorithms()
-        self.update_compute_values(
-            trusted_algorithms, trusted_algorithm_publishers, True, False
-        )
-        assert (
-            self.get_service("compute").compute_values[
-                "publisherTrustedAlgorithmPublishers"
-            ]
-            == trusted_algorithm_publishers
-        ), "New trusted algorithm publisher was not removed. Failed when updating the list of trusted algo publishers. "
-
-        return trusted_algorithm_publishers
-
-    @enforce_types
     def generate_trusted_algorithms(self) -> dict:
         """Returns a trustedAlgorithm dictionary for service at index 0."""
         files = self.get_service_by_index(0).files
@@ -262,50 +203,6 @@ class Asset(AddressCredential):
                 json.dumps(container, separators=(",", ":"))
             ),
         }
-
-    @enforce_types
-    def update_compute_values(
-        self,
-        trusted_algorithms: List,
-        trusted_algo_publishers: Optional[List],
-        allow_network_access: bool,
-        allow_raw_algorithm: bool,
-    ) -> None:
-        """Set the `trusted_algorithms` on the compute service.
-
-        - An assertion is raised if this asset has no compute service
-        - Updates the compute service in place
-        - Adds the trusted algorithms under privacy.publisherTrustedAlgorithms
-
-        :param trusted_algorithms: list of dicts, each dict contain the keys
-            ("containerSectionChecksum", "filesChecksum", "did")
-        :param trusted_algo_publishers: list of strings, addresses of trusted publishers
-        :param allow_network_access: bool -- set to True to allow network access to all the algorithms that belong to this dataset
-        :param allow_raw_algorithm: bool -- determine whether raw algorithms (i.e. unpublished) can be run on this dataset
-        :return: None
-        :raises AssertionError if this asset has no `ServiceTypes.CLOUD_COMPUTE` service
-        """
-        assert not trusted_algorithms or isinstance(trusted_algorithms, list)
-        service = self.get_service(ServiceTypes.CLOUD_COMPUTE)
-        assert service is not None, "this asset does not have a compute service."
-
-        for ta in trusted_algorithms:
-            assert isinstance(
-                ta, dict
-            ), f"item in list of trusted_algorithms must be a dict, got {ta}"
-            assert (
-                "did" in ta
-            ), f"dict in list of trusted_algorithms is expected to have a `did` key, got {ta.keys()}."
-
-        if not service.compute_values:
-            service.compute_values = {}
-
-        service.compute_values["publisherTrustedAlgorithms"] = trusted_algorithms
-        service.compute_values[
-            "publisherTrustedAlgorithmPublishers"
-        ] = trusted_algo_publishers
-        service.compute_values["allowNetworkAccess"] = allow_network_access
-        service.compute_values["allowRawAlgorithm"] = allow_raw_algorithm
 
     @property
     def is_disabled(self) -> bool:
