+++ ocean_lib/assets/trusted_algorithms.py
@@ -90,45 +90,3 @@ def add_publisher_trusted_algorithm_publisher(
     ), "Cannot add trusted algorithm to this asset because it has no compute service."
 
     return compute_service.add_publisher_trusted_algorithm_publisher(publisher_address)
-
-
-@enforce_types
-def remove_publisher_trusted_algorithm(
-    asset_or_did: Union[str, Asset], algo_did: str, metadata_cache_uri: str
-) -> list:
-    """
-    :return: List of trusted algos not containing `algo_did`.
-    """
-    if isinstance(asset_or_did, Asset):
-        asset = asset_or_did
-    else:
-        asset = resolve_asset(asset_or_did, metadata_cache_uri=metadata_cache_uri)
-
-    compute_service = asset.get_service("compute")
-    assert (
-        compute_service
-    ), "Cannot add trusted algorithm to this asset because it has no compute service."
-
-    return asset.remove_publisher_trusted_algorithm(compute_service, algo_did)
-
-
-@enforce_types
-def remove_publisher_trusted_algorithm_publisher(
-    asset_or_did: Union[str, Asset], publisher_address: str, metadata_cache_uri: str
-) -> list:
-    """
-    :return: List of trusted algo publishers not containing `publisher_address`.
-    """
-    if isinstance(asset_or_did, Asset):
-        asset = asset_or_did
-    else:
-        asset = resolve_asset(asset_or_did, metadata_cache_uri=metadata_cache_uri)
-
-    compute_service = asset.get_service("compute")
-    assert (
-        compute_service
-    ), "Cannot add trusted algorithm to this asset because it has no compute service."
-
-    return asset.remove_publisher_trusted_algorithm_publisher(
-        compute_service, publisher_address
-    )
