+++ ocean_lib/services/test/test_service.py
@@ -7,8 +7,15 @@ from unittest.mock import patch
 import pytest
 
 from ocean_lib.assets.asset import Asset
+from ocean_lib.assets.trusted_algorithms import (
+    add_publisher_trusted_algorithm_publisher,
+)
 from ocean_lib.services.service import Service
-from tests.resources.ddo_helpers import get_sample_ddo
+from tests.resources.ddo_helpers import (
+    get_sample_algorithm_ddo,
+    get_sample_ddo,
+    get_sample_ddo_with_compute_service,
+)
 
 
 @pytest.mark.unit
@@ -56,3 +63,99 @@ def test_service():
         assert (
             sa.is_consumable(asset, {"type": "address", "value": "0xdddd"}, True) == 3
         )
+
+
+@pytest.mark.unit
+def test_trusted_algo_functions(publisher_ocean_instance):
+    algorithm_ddo = get_sample_algorithm_ddo(filename="ddo_algorithm2.json")
+    algorithm_ddo.did = "did:op:123"
+    algorithm_ddo_v2 = get_sample_algorithm_ddo(filename="ddo_algorithm2.json")
+    algorithm_ddo_v2.did = "did:op:1234"
+    algorithm_ddo_v3 = get_sample_algorithm_ddo(filename="ddo_algorithm2.json")
+    algorithm_ddo_v3.did = "did:op:3333"
+
+    ddo_dict = get_sample_ddo_with_compute_service()
+    service_dict = ddo_dict["services"][1]
+    compute_service = Service.from_dict(service_dict)
+    assert compute_service.type == "compute"
+
+    # remove an existing algorithm to publisher_trusted_algorithms list
+    new_publisher_trusted_algorithms = (
+        compute_service.remove_publisher_trusted_algorithm(algorithm_ddo.did)
+    )
+
+    assert (
+        new_publisher_trusted_algorithms is not None
+    ), "Remove process of a trusted algorithm failed."
+    assert len(new_publisher_trusted_algorithms) == 1
+
+    # remove a trusted algorithm that does not belong to publisher_trusted_algorithms list
+    new_publisher_trusted_algorithms = (
+        compute_service.remove_publisher_trusted_algorithm(algorithm_ddo_v3.did)
+    )
+    assert len(new_publisher_trusted_algorithms) == 1
+
+
+@pytest.mark.unit
+def test_utilitary_functions_for_trusted_algorithm_publishers(publisher_ocean_instance):
+    """Tests adding/removing trusted algorithms in the DDO metadata."""
+    ddo = Asset.from_dict(get_sample_ddo_with_compute_service())
+    compute_service = ddo.services[1]
+    assert compute_service.type == "compute"
+
+    addr1 = publisher_ocean_instance.web3.eth.account.create().address
+    compute_service.compute_values["publisherTrustedAlgorithmPublishers"] = [addr1]
+
+    addr2 = publisher_ocean_instance.web3.eth.account.create().address
+    # add a new trusted algorithm to the publisher_trusted_algorithms list
+    new_publisher_trusted_algo_publishers = add_publisher_trusted_algorithm_publisher(
+        ddo, addr2, publisher_ocean_instance.config.metadata_cache_uri
+    )
+
+    assert (
+        new_publisher_trusted_algo_publishers is not None
+    ), "Added a new trusted algorithm failed. The list is empty."
+    assert len(new_publisher_trusted_algo_publishers) == 2
+
+    # add an existing algorithm to publisher_trusted_algorithms list
+    new_publisher_trusted_algo_publishers = add_publisher_trusted_algorithm_publisher(
+        ddo, addr2.upper(), publisher_ocean_instance.config.metadata_cache_uri
+    )
+    assert len(new_publisher_trusted_algo_publishers) == 2
+
+    # remove an existing algorithm to publisher_trusted_algorithms list
+    new_publisher_trusted_algo_publishers = (
+        compute_service.remove_publisher_trusted_algorithm_publisher(addr2.upper())
+    )
+
+    assert len(new_publisher_trusted_algo_publishers) == 1
+
+    addr3 = publisher_ocean_instance.web3.eth.account.create().address
+    # remove a trusted algorithm that does not belong to publisher_trusted_algorithms list
+    new_publisher_trusted_algo_publishers = (
+        compute_service.remove_publisher_trusted_algorithm_publisher(addr3)
+    )
+    assert len(new_publisher_trusted_algo_publishers) == 1
+
+
+@pytest.mark.unit
+def test_inexistent_removals():
+    ddo_dict = get_sample_ddo_with_compute_service()
+    del ddo_dict["services"][1]["compute"]["publisherTrustedAlgorithms"]
+    ddo = Asset.from_dict(ddo_dict)
+    compute_service = ddo.get_service("compute")
+
+    with pytest.raises(
+        ValueError, match="Algorithm notadid is not in trusted algorithms"
+    ):
+        compute_service.remove_publisher_trusted_algorithm("notadid")
+
+    ddo_dict = get_sample_ddo_with_compute_service()
+    del ddo_dict["services"][1]["compute"]["publisherTrustedAlgorithmPublishers"]
+    ddo = Asset.from_dict(ddo_dict)
+    compute_service = ddo.get_service("compute")
+
+    with pytest.raises(
+        ValueError, match="Publisher notadid is not in trusted algorithm publishers"
+    ):
+        compute_service.remove_publisher_trusted_algorithm_publisher("notadid")
